+++ CJBCheatsMenu/release-notes.md
@@ -3,6 +3,8 @@
 # Release notes
 ## Upcoming release
 * Added support for setting relationships for unmet villagers.
+* Fast machine processing is now much faster.
+* Fast machine processing now continues working when time is paused.
 * Fast machine list is now sorted by name.
 * Updated translations. Thanks to kelvindules (updated Portuguese) and VincentRoth (added French)!